Defer use of external dependencies in setup.py.
This commit adds a BioCythonNumpyBuildExt class, which waits until dependencies required for the setup were added and then uses the necessary extensions to

- cythonize the tn93 extension,
- get the numpy include directory (also bumb np to v1.22.x), and
- add the package data (references need biopython).

Not quite sure if this couldn't have been done a little nicer in setup.cfg, but the build time deps are encapsulated at least.

This approach is actually already kind of deprecated, but the better (or at least more modern) approach is to conform to PEP-517,
which would probably necessitate migrating some of the project config to a ``pyproject.toml`` file.
Details regarding the newer approach can be found here https://peps.python.org/pep-0517/